refactor: Move standard library to own directory
The standard library implementation was previously nested within the
compiler, which made it difficult to maintain a clear separation between
the core language and the standard library. Hence, this patch moves the
library to its own directory. Further, the documentation-related code is
split off from module handling.
