+++ packages/language/src/compiler/generator.ts
@@ -3,6 +3,7 @@ import type { Bus, BusId, InstrumentId, InstrumentRouting, Mixer, MixerRouting, 
 import { concatPatterns, createParallelPattern, createSerialPattern, mergePatterns, multiplyPattern } from '@core'
 import type { Numeric, Unit } from '@utility'
 import { numeric } from '@utility'
+import { getStandardModuleValue } from '../library/modules.js'
 import type { Function } from '../type-system/base/function.js'
 import { FunctionFacet } from '../type-system/base/function.js'
 import { ModuleFacet } from '../type-system/base/module.js'
@@ -23,11 +24,10 @@ import { busSchema, partSchema, stepSchema, trackSchema } from './common.js'
 import type { CurveSegment as GeneratedCurveSegment } from './curves.js'
 import { createCurve, createCurveSegment, getCurveSegmentType, renderCurvePoints } from './curves.js'
 import { CompileError } from './error.js'
-import { getStandardModuleValue } from './modules.js'
 import type { GenerateOptions } from './options.js'
 import type { GlobalScope, MutableScope, Scope } from './scopes.js'
 import { createGlobalScope, createLocalScope, resolveInScope } from './scopes.js'
-import { BusType, Numbers, Parameters } from './type-helpers.js'
+import { BusType, Numbers, Parameters } from '../type-system/helpers.js'
 import { isSyntaxUnit, toNumberValue } from './units.js'
 
 /**
