+++ packages/language/src/library/documentation.ts
@@ -1,41 +1,9 @@
-import type { Module } from '../type-system/base/module.js'
 import type { Function } from '../type-system/base/function.js'
 import { FunctionFacet } from '../type-system/base/function.js'
+import type { Module } from '../type-system/base/module.js'
 import { ModuleFacet } from '../type-system/base/module.js'
 import type { Value } from '../type-system/types.js'
-import { effectsModule } from './modules/effects.js'
-import { instrumentsModule } from './modules/instruments.js'
-import { patternsModule } from './modules/patterns.js'
-
-const standardLibrary = Object.freeze(new Map([
-  ['patterns', patternsModule],
-  ['instruments', instrumentsModule],
-  ['effects', effectsModule]
-]))
-
-const standardLibraryModuleNames: ReadonlySet<string> = Object.freeze(new Set(standardLibrary.keys()))
-
-export function getStandardModuleNames (): ReadonlySet<string> {
-  return standardLibraryModuleNames
-}
-
-export function getStandardModuleValue (name: string): Value<typeof ModuleFacet> | undefined {
-  const value = standardLibrary.get(name)
-  if (value == null) {
-    return undefined
-  }
-
-  return value as Value<typeof ModuleFacet>
-}
-
-export function getStandardModule (name: string): Module | undefined {
-  const value = getStandardModuleValue(name)
-  if (value == null) {
-    return undefined
-  }
-
-  return ModuleFacet.get(value)
-}
+import { getStandardModule } from './modules.js'
 
 export interface Documentation {
   readonly title: string
