+++ packages/language/src/library/modules/instruments.ts
@@ -1,15 +1,15 @@
 import type { Envelope, Oscillator } from '@core'
 import { isPitch } from '@core'
 import { numeric } from '@utility'
+import type { InstrumentContext, ParameterContext } from '../../compiler/scopes.js'
 import { NumberFacet } from '../../type-system/base/number.js'
 import { RecordFacet } from '../../type-system/base/record.js'
 import { StringFacet } from '../../type-system/base/string.js'
 import { InstrumentFacet } from '../../type-system/domain/instrument.js'
 import { ParameterFacet } from '../../type-system/domain/parameter.js'
 import { makeType } from '../../type-system/factory.js'
+import { Functions, Modules, Parameters } from '../../type-system/helpers.js'
 import type { Value } from '../../type-system/types.js'
-import type { InstrumentContext, ParameterContext } from '../scopes.js'
-import { Functions, Modules, Parameters } from '../type-helpers.js'
 
 const UNITY_GAIN = numeric('db', 0)
