+++ packages/language/src/type-system/helpers.ts
@@ -1,16 +1,16 @@
 import type { Parameter } from '@core'
 import type { Numeric, Unit } from '@utility'
-import type { Function } from '../type-system/base/function.js'
-import { FunctionFacet } from '../type-system/base/function.js'
-import type { Module } from '../type-system/base/module.js'
-import { ModuleFacet } from '../type-system/base/module.js'
-import { NumberFacet } from '../type-system/base/number.js'
-import { RecordFacet } from '../type-system/base/record.js'
-import { BusFacet } from '../type-system/domain/bus.js'
-import { ParameterFacet } from '../type-system/domain/parameter.js'
-import { makeType } from '../type-system/factory.js'
-import type { Schema } from '../type-system/schema.js'
-import type { Facet, FacetType, Value } from '../type-system/types.js'
+import type { Function } from './base/function.js'
+import { FunctionFacet } from './base/function.js'
+import type { Module } from './base/module.js'
+import { ModuleFacet } from './base/module.js'
+import { NumberFacet } from './base/number.js'
+import { RecordFacet } from './base/record.js'
+import { BusFacet } from './domain/bus.js'
+import { ParameterFacet } from './domain/parameter.js'
+import { makeType } from './factory.js'
+import type { Schema } from './schema.js'
+import type { Facet, FacetType, Value } from './types.js'
 
 export const Functions = {
   of: <const S extends Schema, const R extends FacetType, const Context> (value: Function<S, R, Context>): Value => {
